move libvirt backend to own dir
