+++ src/backend_pool/libvirt/snapshot_handler.py
